Move alert components to UI namespace, drop TimeLocalizer preview
- Rename Alerts::FlashMessages -> UI::Alerts::FlashMessage and
  Alerts::ObjectErrors -> UI::Alerts::ObjectError (files, i18n scope,
  and all view/spec references); carry existing nl/it/es/nb
  translations to the new scope.
- Remove the preview-only UI::TimeLocalizer component and its system
  spec; the live time_localizer.js functionality is untouched.
- Group all DefinitionList::Container previews under one "Terms" group
  and drop "redesign" wording from the component/preview.
