--- app/components/alerts/object_errors/component.en.yml
@@ -1,7 +1,0 @@
----
-en:
-  components:
-    alerts:
-      object_errors:
-        errors_prevented_this_from_being_saved: "%{errors_count} prevented this %{object_name}
-          from being saved:"
