+++ app/components/ui/alerts/object_error/component.en.yml
@@ -1,0 +1,8 @@
+---
+en:
+  components:
+    ui:
+      alerts:
+        object_error:
+          errors_prevented_this_from_being_saved: "%{errors_count} prevented this %{object_name}
+            from being saved:"
