+++ app/components/ui/alerts/object_error/component.html.erb
