+++ app/components/ui/alerts/object_error/component_preview.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module UI
+  module Alerts
+    module ObjectError
+      class ComponentPreview < ApplicationComponentPreview
+        def default
+          user = User.new
+          user.errors.add(:email, "can't be blank")
+          user.errors.add(:password, "is too short")
+          render(UI::Alerts::ObjectError::Component.new(object: user))
+        end
+      end
+    end
+  end
+end
