--- app/components/ui/time_localizer/component_preview/default.html.erb
@@ -1,86 +1,0 @@
-<%# NOTE: this component is here for testing. It uses the same time_localizer.js setup as everything else %>
-
-<% time_zone ||= nil %>
-
-<% if time_zone.present? %>
-  <script>
-    window.localTimezone = <%== time_zone.to_json %>;
-  </script>
-<% end %>
-
-<h2 class="tw:mb-4 tw:text-sm">
-  <span class="tw:opacity-65">In time zone:</span>
-  <span class="localizeTimezone"></span>
-</h2>
-
-<div class="tw:mb-4">
-  <p>
-    <span class="tw:text-sm tw:opacity-65">Current time:</span>
-
-    <span class="localizeTime">
-      <%= l ::Time.current, format: :convert_time %>
-    </span>
-  </p>
-</div>
-
-<div class="tw:mb-4">
-  <p>
-    <span class="tw:text-sm tw:opacity-65">Yesterday:</span>
-
-    <span class="localizeTime">
-      <%= l (::Time.current - 1.day), format: :convert_time %>
-    </span>
-  </p>
-</div>
-
-<div class="tw:mb-4">
-  <p>
-    <span class="tw:text-sm tw:opacity-65">Tomorrow:</span>
-
-    <span class="localizeTime">
-      <%= l (::Time.current + 1.day), format: :convert_time %>
-    </span>
-  </p>
-</div>
-
-<div class="tw:mb-4">
-  <p>
-    <span class="tw:text-sm tw:opacity-65">One week ago:</span>
-
-    <span class="localizeTime">
-      <%= l (::Time.current - 7.days), format: :convert_time %>
-    </span>
-  </p>
-</div>
-
-<div class="tw:mb-4">
-  <p>
-    <span class="tw:text-sm tw:opacity-65">One year ago:</span>
-
-    <span class="localizeTime">
-      <%= l (::Time.current - 1.year), format: :convert_time %>
-    </span>
-  </p>
-</div>
-
-<div class="tw:mb-4">
-  <p>
-    <span class="tw:text-sm tw:opacity-65">Yesterday (precise time):</span>
-
-    <span class="localizeTime preciseTime">
-      <%= l (::Time.current - 1.day), format: :convert_time %>
-    </span>
-  </p>
-</div>
-
-<div class="tw:mb-4">
-  <p>
-    <span class="tw:text-sm tw:opacity-65">
-      One year ago (precise time seconds):
-    </span>
-
-    <span class="localizeTime preciseTimeSeconds">
-      <%= l (::Time.current - 1.year), format: :convert_time %>
-    </span>
-  </p>
-</div>
