+++ app/views/admin/ads/_form.html.haml
@@ -1,4 +1,4 @@
-= render(Alerts::ObjectErrors::Component.new(object: @ad))
+= render(UI::Alerts::ObjectError::Component.new(object: @ad))
 
 .form-group
   = f.label :organization
