+++ app/views/admin/bike_stickers/new.html.haml
@@ -51,7 +51,7 @@
 
 
 = form_for [:admin, @bike_sticker_batch], url: {action: "create", controller: "bike_stickers"} do |f|
-  = render(Alerts::ObjectErrors::Component.new(name: "Sticker Batch", object: @bike_sticker_batch))
+  = render(UI::Alerts::ObjectError::Component.new(name: "Sticker Batch", object: @bike_sticker_batch))
 
   .row.mt-4
     .col-md-6.col-lg-3
