+++ app/views/admin/bikes/edit.html.haml
@@ -11,7 +11,7 @@
     = render partial: "/bikes/credibility_score", locals: { bike: @bike }
 
 = form_for [:admin, @bike], html: { class: 'form form-horizontal admin-bike-edit' } do |f|
-  = render(Alerts::ObjectErrors::Component.new(object: @bike))
+  = render(UI::Alerts::ObjectError::Component.new(object: @bike))
 
   .row
     .col-md-4.col-lg-3
