+++ app/views/admin/bulk_imports/new.html.haml
@@ -5,7 +5,7 @@
 .row
   .col-sm-6
     = form_for [:admin, @bulk_import] do |f|
-      = render(Alerts::ObjectErrors::Component.new(object: @bulk_import))
+      = render(UI::Alerts::ObjectError::Component.new(object: @bulk_import))
 
       .form-group
         = render Form::Combobox::Component.new(name: :organization_id, form: f, label: BulkImport.human_attribute_name(:organization_id), placeholder: "Choose organization", required: true, include_blank: true, options: Organization.all.pluck(:name, :id))
