+++ app/views/admin/content_tags/_form.html.haml
@@ -1,5 +1,5 @@
 = form_for [:admin, @content_tag] do |f|
-  = render(Alerts::ObjectErrors::Component.new(object: @content_tag))
+  = render(UI::Alerts::ObjectError::Component.new(object: @content_tag))
 
   .row
     .col-md-6
