+++ app/views/admin/memberships/_form.html.haml
@@ -2,7 +2,7 @@
   .card-body
     = form_for [:admin, @membership] do |f|
       - if @membership.errors.any?
-        = render(Alerts::ObjectErrors::Component.new(object: @membership, dismissable: true))
+        = render(UI::Alerts::ObjectError::Component.new(object: @membership, dismissable: true))
 
       - if @membership.id.blank?
         .form-group
