+++ app/views/admin/news/edit.html.haml
@@ -22,7 +22,7 @@
 
 - url = {action: "update", controller: "news"}
 = form_for [:admin, @blog], :url => url do |f|
-  = render(Alerts::ObjectErrors::Component.new(object: @blog))
+  = render(UI::Alerts::ObjectError::Component.new(object: @blog))
 
   - if @blog.is_listicle
     %h2.text-danger
