+++ app/views/admin/organization_features/_form.html.haml
@@ -1,5 +1,5 @@
 = form_for [:admin, @organization_feature] do |f|
-  = render(Alerts::ObjectErrors::Component.new(object: @organization_feature))
+  = render(UI::Alerts::ObjectError::Component.new(object: @organization_feature))
 
   .row
     .col-md-6.col-xs-12
