+++ app/views/admin/organization_roles/_form.html.haml
@@ -7,7 +7,7 @@
           %br
           It would be better to delete and re-add
     = form_for [:admin, @organization_role] do |f|
-      = render(Alerts::ObjectErrors::Component.new(object: @organization_role))
+      = render(UI::Alerts::ObjectError::Component.new(object: @organization_role))
 
       .form-group
         = render Form::Combobox::Component.new(name: :organization_id, form: f, label: OrganizationRole.human_attribute_name(:organization), placeholder: "Please select an organization", required: true, include_blank: true, options: @organizations.pluck(:name, :id))
