+++ app/views/admin/organizations/_form.html.haml
@@ -1,4 +1,4 @@
-= render(Alerts::ObjectErrors::Component.new(object: @organization))
+= render(UI::Alerts::ObjectError::Component.new(object: @organization))
 
 .row
   .form-group.col-md-6
