+++ app/views/admin/organizations/custom_layouts/_mail_snippet.html.haml
@@ -1,5 +1,5 @@
 .form.form-horizontal.clearfix
-  = render(Alerts::ObjectErrors::Component.new(object: @organization))
+  = render(UI::Alerts::ObjectError::Component.new(object: @organization))
 
   = form_for @organization, {as: :organization, url: admin_organization_custom_layout_path(organization_id: @organization.to_param, id: params[:id]), action: 'update'} do |f|
     = f.fields_for :mail_snippets do |snippet|
