+++ app/views/admin/organizations/invoices/_form.html.haml
@@ -1,7 +1,7 @@
 #invoiceForm.mt-4
   = form_for [:admin, @organization, @invoice] do |f|
     = f.hidden_field :organization_feature_ids
-    = render(Alerts::ObjectErrors::Component.new(object: @invoice))
+    = render(UI::Alerts::ObjectError::Component.new(object: @invoice))
 
     .full-screen-table
       %table.table.table-striped.table-bordered.table-sm
