+++ app/views/admin/ownerships/edit.html.haml
@@ -16,7 +16,7 @@
 .row
   .col-md-6
     = form_for [:admin, @ownership], html: { class: "form form-horizontal" } do |f|
-      = render(Alerts::ObjectErrors::Component.new(object: @ownership))
+      = render(UI::Alerts::ObjectError::Component.new(object: @ownership))
 
       .form-group
         = f.label :creator_id
