+++ app/views/admin/paints/edit.html.haml
@@ -32,7 +32,7 @@
 
 
 = form_for [:admin, @paint] do |f|
-  = render(Alerts::ObjectErrors::Component.new(object: @paint))
+  = render(UI::Alerts::ObjectError::Component.new(object: @paint))
 
   .row
     .col-lg-3.col-md-6.col-sm-12
