+++ app/views/admin/payments/edit.html.haml
@@ -20,7 +20,7 @@
   .row
     .col-md-8.offset-md-2
       = form_for [:admin, @payment] do |f|
-        = render(Alerts::ObjectErrors::Component.new(object: @payment))
+        = render(UI::Alerts::ObjectError::Component.new(object: @payment))
 
         .row
           .col-lg-6
