+++ app/views/admin/recovery_displays/_form.html.erb
@@ -62,7 +62,7 @@
     "admin--recovery-display-form-max-character-count-value" => Integrations::SocialPoster::TWEET_LENGTH,
     "admin--recovery-display-form-remote-photo-url-value" => @bike&.image_url || ""
   }} do |f| %>
-  <%= render(Alerts::ObjectErrors::Component.new(object: @recovery_display, dismissable: true)) %>
+  <%= render(UI::Alerts::ObjectError::Component.new(object: @recovery_display, dismissable: true)) %>
 
   <% if @recovery_display.stolen_record_id.present? %>
     <%= f.hidden_field :stolen_record_id, value: @recovery_display.stolen_record_id %>
