+++ app/views/admin/social_posts/new.html.haml
@@ -4,7 +4,7 @@
       New Post
 
 #postForm
-  = render(Alerts::ObjectErrors::Component.new(object: @social_post))
+  = render(UI::Alerts::ObjectError::Component.new(object: @social_post))
 
   = form_for [:admin, @social_post], url: { action: "create", controller: "social_posts" } do |f|
     .row
