+++ app/views/admin/stolen_bikes/edit.html.haml
@@ -94,7 +94,7 @@
       Send bike owner an email
 
     - if @customer_contact.errors.any?
-      = render(Alerts::ObjectErrors::Component.new(object: @customer_contact))
+      = render(UI::Alerts::ObjectError::Component.new(object: @customer_contact))
     .row
       .col-lg-4
         .row
