+++ app/views/admin/superuser_abilities/edit.html.haml
@@ -45,7 +45,7 @@
     .card.bg-light
       .card-body
         = form_for [:admin, @superuser_ability] do |f|
-          = render(Alerts::ObjectErrors::Component.new(object: @superuser_ability))
+          = render(UI::Alerts::ObjectError::Component.new(object: @superuser_ability))
 
           - SuperuserAbility::SU_OPTIONS.each do |ability|
             .form-group
