+++ app/views/my_accounts/messages/_form.html.erb
@@ -5,7 +5,7 @@
 </h2>
 
 <%= form_with(model: marketplace_message, url: my_account_messages_path, method: :post, local: true, class: "") do |f| %>
-  <%= render(Alerts::ObjectErrors::Component.new(object: marketplace_message, name: t(".message"), dismissable: true)) %>
+  <%= render(UI::Alerts::ObjectError::Component.new(object: marketplace_message, name: t(".message"), dismissable: true)) %>
 
   <%= f.hidden_field :initial_record_id %>
   <%= f.hidden_field :marketplace_listing_id %>
