+++ app/views/organized/users/new.html.haml
@@ -20,7 +20,7 @@
 
 = form_for @organization_role, { as: :organization_role, url: organization_users_path(organization_id: current_organization.to_param), action: 'create', html: { class: 'organized-form' } } do |f|
 
-  = render(Alerts::ObjectErrors::Component.new(object: @organization_role, name: t(".organization_role")))
+  = render(UI::Alerts::ObjectError::Component.new(object: @organization_role, name: t(".organization_role")))
 
   - display_multi_invite = params[:multiple_emails_invited].present?
   .form-group.row#singleEmailInvite{ class: display_multi_invite ? "currently-hidden" : "" }
