+++ config/locales/translation.it.yml
@@ -6037,10 +6037,10 @@ it:
         serial_label: 'Seriale:'
         view_all: visualizza tutto
         sticker_label: 'Adesivo:'
-    alerts:
-      object_errors:
-        errors_prevented_this_from_being_saved: "%{errors_count} ha impedito il salvataggio
-          di questo %{object_name}:"
+      alerts:
+        object_error:
+          errors_prevented_this_from_being_saved: "%{errors_count} ha impedito il salvataggio
+            di questo %{object_name}:"
     form:
       legacy_form_well:
         address_record:
