+++ config/locales/translation.nb.yml
@@ -5763,10 +5763,10 @@ nb:
         serial_label: 'Serie:'
         view_all: se alle
         sticker_label: 'Klistremerke:'
-    alerts:
-      object_errors:
-        errors_prevented_this_from_being_saved: "%{errors_count} forhindret at denne
-          %{object_name} ble lagret:"
+      alerts:
+        object_error:
+          errors_prevented_this_from_being_saved: "%{errors_count} forhindret at denne
+            %{object_name} ble lagret:"
     form:
       legacy_form_well:
         address_record:
